requirements.txt
sentence-transformers==2.2.2
groq==0.4.0
python-dotenv==1.0.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.0
pandas==2.1.0
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.0
requests==2.31.0
httpx==0.24.1 